+++ src/pages/auth/utils/passwordStrength.ts
